backport: Permissions improvements
